requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
idna==2.10
lxml==4.6.2
python-dateutil==2.8.1
requests==2.25.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.26.2
